scripts/requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
idna==3.10
lark==1.2.2
python-hcl2==6.1.1
PyYAML==6.0.2
requests==2.32.3
urllib3==2.2.3
